Moved Gig categories to Gig.cs
